% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Characteristic_Species}
\alias{Characteristic_Species}
\title{Characteristic species for different habitat types}
\format{
A data frame with 847 rows and 16 columns:
}
\usage{
Characteristic_Species
}
\description{
Characteristic species for different habitat types
}
\keyword{datasets}
