% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Ellenberg_CSR}
\alias{Ellenberg_CSR}
\title{CSR values and Ellenberg values for different species}
\format{
A data frame with 2,827 rows and 36 columns:
}
\usage{
Ellenberg_CSR
}
\description{
CSR values and Ellenberg values for different species
}
\keyword{datasets}
