% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Final_Frequency}
\alias{Final_Frequency}
\title{A subset of data from from a sampled dataset with the sampled species}
\format{
A data frame with 723,101 rows and 5 columns:
\describe{
  \item{plot}{Code for the plot}
  \item{year}{The sampled year}
  \item{canonicalName}{name of the species subspecies or variety}
  \item{species}{name of the species, solved for synonyms}
  \item{rank}{Taxonomic level it has been solved to
}}
}
\usage{
Final_Frequency
}
\description{
A subset of data from from a sampled dataset with the sampled species
}
\keyword{datasets}
