% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Final_Frequency}
\alias{Final_Frequency}
\title{A subset of data from from a sampled dataset with the sampled species with added taxon_id_Arter and photo_file.}
\format{
A data frame with 615860 rows and 6 columns:
\describe{
  \item{Taxa}{name of the species subspecies or variety}
  \item{plot}{Code for the plot}
  \item{year}{The sampled year}
  \item{species}{name of the species, solved for synonyms}
  \item{taxon_id_Arter}{taxon_id for linking to Arter.dk}
  \item{photo_file}{file name for photos in folder inst/Pictures}
}
}
\usage{
Final_Frequency
}
\description{
A subset of data from from a sampled dataset with the sampled species with added taxon_id_Arter and photo_file.
}
\keyword{datasets}
