% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{SpatialData}
\alias{SpatialData}
\title{A subset of data from a sampled dataset with habitat types and plot code
of sampling.}
\format{
## A data frame with 46,536 rows and 5 columns:
\describe{
  \item{plot}{Code for the plot}
  \item{habtype}{Code for habitat type}
  \item{MajorHab}{Code for the major habitat type}
  \item{habitat_name}{Name for the habitat type}
  \item{MajorHabName}{Name for the major habitat type}
}
}
\usage{
SpatialData
}
\description{
A subset of data from a sampled dataset with habitat types and plot code
of sampling.
}
\keyword{datasets}
